Extracting `DateTimePicker` component.
